Add note about GH Pages redeploy
I changed `master` to `main` on a `username.github.io` repo and the site went to 404. I assume it was DNS related and would solve itself and went to sleep. The site was still down in the morning, so I edited the `index.html` to add a single character change and merged to trigger a deploy. This fixed it. I'm shocked that GitHub doesn't just perform a one-time deploy of GH Pages after a branch name. Especially in a scenario where you're going from `master` to `main` on a repo ending in `.github.io`. But really it should just do that anytime the branch used for GH Pages is renamed.